+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 UNRELEASED
 ----------
 
+* [ [#1808](https://github.com/digitalfabrik/integreat-cms/issues/1808) ] Improve calculation of HIX values via Textlab
 * [ [#1800](https://github.com/digitalfabrik/integreat-cms/issues/1800) ] Exclude archived pages from PDF exports
 * [ [#1802](https://github.com/digitalfabrik/integreat-cms/issues/1802) ] Reenable table of contents and page numbers in PDFs
 * [ [#1350](https://github.com/digitalfabrik/integreat-cms/issues/1350) ] Various small PDF export improvements